split solution into two projects: UI and class library
