+++ VaccinationLibrary/Class1.cs
@@ -1,0 +1,7 @@
+﻿namespace VaccinationPassportLibrary
+{
+    public class Class1
+    {
+
+    }
+}
