+++ VaccinationLibrary/Models/Person.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Vaccination_passport.Models
+namespace VaccinationPassportLibrary.Models
 {
     public class Person
     {
