+++ Vaccination_passport/App.xaml
@@ -1,8 +1,8 @@
 ﻿<Application
-    x:Class="Vaccination_passport.App"
+    x:Class="VaccinationLibrary.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="clr-namespace:Vaccination_passport"
+    xmlns:local="clr-namespace:VaccinationLibrary"
     StartupUri="Windows/MainWindow.xaml">
     <Application.Resources>
         <ResourceDictionary>
