+++ Vaccination_passport/App.xaml.cs
@@ -1,6 +1,6 @@
 ﻿using System.Windows;
 
-namespace Vaccination_passport
+namespace VaccinationLibrary
 {
     /// <summary>
     /// Interaction logic for App.xaml
