+++ Vaccination_passport/Resources/DB.xaml
@@ -1,6 +1,6 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:db="clr-namespace:Vaccination_passport.Models">
+                    xmlns:db="clr-namespace:VaccinationLibrary.Models">
 
     <db:DataAccess x:Key="SuperDB"></db:DataAccess>
