+++ Vaccination_passport/UserControls/PersonDataBlock.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="Vaccination_passport.User_Controls.PersonDataBlock"
+﻿<UserControl x:Class="VaccinationLibrary.User_Controls.PersonDataBlock"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:Vaccination_passport.User_Controls"
+             xmlns:local="clr-namespace:VaccinationLibrary.User_Controls"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="799">
     <Grid>
