+++ Vaccination_passport/UserControls/PersonDataBlock.xaml.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Windows.Controls;
 
-namespace Vaccination_passport.User_Controls
+namespace VaccinationLibrary.User_Controls
 {
     /// <summary>
     /// Interaction logic for PersonDataBlock.xaml
