+++ Vaccination_passport/VaccinationPassportUI.csproj
@@ -14,7 +14,10 @@
 
   <ItemGroup>
     <PackageReference Include="Extended.Wpf.Toolkit" Version="4.1.0" />
-    <PackageReference Include="System.Data.OleDb" Version="6.0.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\VaccinationLibrary\VaccinationPassportLibrary.csproj" />
   </ItemGroup>
 
   <ItemGroup>
