+++ Vaccination_passport/Windows/MainWindow.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="Vaccination_passport.Windows.MainWindow"
+﻿<Window x:Class="VaccinationLibrary.Windows.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Vaccination_passport"
+        xmlns:local="clr-namespace:VaccinationLibrary"
         xmlns:tk ="http://schemas.xceed.com/wpf/xaml/toolkit"
         mc:Ignorable="d"
         Title="MainWindow" Height="450" Width="800" Icon="{StaticResource Logo}">
