+++ Vaccination_passport/Windows/NewPassWindow.xaml.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using System.Windows;
-using Vaccination_passport.Models;
+using VaccinationPassportUI.Models;
 
-namespace Vaccination_passport.Windows
+namespace VaccinationLibrary.Windows
 {
     /// <summary>
     /// Interaction logic for NewPassWindow.xaml
