+++ Vaccination_passport/Windows/PassportWindow.xaml
@@ -1,10 +1,10 @@
-﻿<Window x:Class="Vaccination_passport.Windows.PassportWindow"
+﻿<Window x:Class="VaccinationLibrary.Windows.PassportWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Vaccination_passport.Windows"
-        xmlns:uc="clr-namespace:Vaccination_passport.User_Controls"
+        xmlns:local="clr-namespace:VaccinationLibrary.Windows"
+        xmlns:uc="clr-namespace:VaccinationLibrary.User_Controls"
         mc:Ignorable="d"
         DataContext="{DynamicResource currentPerson}"
         Title="PassportWindow" Height="450" Width="800" Icon="{StaticResource Logo}">
